+++ internal/cloud/azure/BUILD.bazel
@@ -7,6 +7,8 @@ go_library(
         "azure.go",
         "imds.go",
         "interface.go",
+        "iptables_cross.go",
+        "iptables_linux.go",
     ],
     importpath = "github.com/edgelesssys/constellation/v2/internal/cloud/azure",
     visibility = ["//:__subpackages__"],
@@ -20,9 +22,17 @@ go_library(
         "@com_github_azure_azure_sdk_for_go_sdk_azidentity//:azidentity",
         "@com_github_azure_azure_sdk_for_go_sdk_resourcemanager_compute_armcompute_v5//:armcompute",
         "@com_github_azure_azure_sdk_for_go_sdk_resourcemanager_network_armnetwork_v5//:armnetwork",
-        "@io_k8s_kubernetes//pkg/util/iptables",
-        "@io_k8s_utils//exec",
-    ],
+    ] + select({
+        "@io_bazel_rules_go//go/platform:android": [
+            "@io_k8s_kubernetes//pkg/util/iptables",
+            "@io_k8s_utils//exec",
+        ],
+        "@io_bazel_rules_go//go/platform:linux": [
+            "@io_k8s_kubernetes//pkg/util/iptables",
+            "@io_k8s_utils//exec",
+        ],
+        "//conditions:default": [],
+    }),
 )
 
 go_test(
